+++ tests/test_core.py
@@ -266,6 +266,33 @@ def test_remove_variable():
     assert expected == res
 
 
+def test_add_aggregation():
+    nc = xncml.Dataset(input_file)
+    nc.add_aggregation('new_dim', 'joinNew')
+    nc.add_variable_agg('new_dim', 'newVar')
+
+    expected = [OrderedDict([('@dimName', 'new_dim'), ('@type', 'joinNew'),
+                             ('variableAgg', [OrderedDict([('@name', 'newVar')])])])]
+    res = nc.ncroot['netcdf']['aggregation']
+
+    assert expected == res
+
+
+def test_add_scan():
+    nc = xncml.Dataset(input_file)
+    nc.add_aggregation('new_dim', 'joinExisting')
+    nc.add_scan('new_dim', location='foo', suffix='.nc')
+
+    expected = [OrderedDict([('@dimName', 'new_dim'), ('@type', 'joinExisting'),
+                                ('scan', [OrderedDict([
+                                    ('@location', 'foo'),
+                                    ('@subdirs', 'true'),
+                                    ('@suffix', '.nc')])])])]
+
+    res = nc.ncroot['netcdf']['aggregation']
+    assert expected == res
+
+
 def test_to_ncml():
     nc = xncml.Dataset(input_file)
     with tempfile.NamedTemporaryFile(suffix='.ncml') as t: